requirements.txt
ollama
llama-index
EbookLib
html2text
llama-index-embeddings-huggingface
llama-index-llms-groq
llama-index-embeddings-openai

streamlit
langchain
langchain-community
langchain_openai
langchain-groq
langchain_core
openai
groq
pypdf
PyPDF2
python-dotenv
faiss-cpu
bs4